+++ xrpl/core/binarycodec/__init__.py
@@ -2,10 +2,12 @@
 Functions for encoding objects into the XRP Ledger's canonical
 binary format and decoding them.
 """
+
 from xrpl.core.binarycodec.exceptions import XRPLBinaryCodecException
 from xrpl.core.binarycodec.main import (
     decode,
     encode,
+    encode_for_batch,
     encode_for_multisigning,
     encode_for_signing,
     encode_for_signing_claim,
@@ -14,6 +16,7 @@ from xrpl.core.binarycodec.main import (
 __all__ = [
     "decode",
     "encode",
+    "encode_for_batch",
     "encode_for_multisigning",
     "encode_for_signing",
     "encode_for_signing_claim",